@@ -198,7 +198,7 @@ String simplifyHashCode(String str) {
   });
 }
 
-Finder findText(String data) => _RichTextFinder(data);
+Finder findText(String data) => _TextFinder(data);
 
 Future<int> tapText(WidgetTester tester, String data) async {
   var tapped = 0;
@@ -655,46 +655,52 @@ class Explainer {
       attr.add('message=${widget.message}');
     }
 
-    // A-F
-    // `RichText` is an exception, it is a `MultiChildRenderObjectWidget`
-    // so it has to be processed first
+    // Special cases
+    // `RichText` is a `MultiChildRenderObjectWidget` but needs different logic
     attr.add(
       widget is RichText
           ? _inlineSpan(widget.text)
-          : widget is Container
+          : widget is MultiChildRenderObjectWidget
+              ? _widgetChildren(widget.children)
+              : '',
+    );
+    // `MouseRegion` is a `SingleChildRenderObjectWidget` since Flutter 2.11
+    // (see https://github.com/flutter/flutter/pull/96636)
+    attr.add(
+      widget is MouseRegion
+          ? _widgetChild(widget.child)
+          : widget is SingleChildRenderObjectWidget
               ? _widgetChild(widget.child)
               : '',
     );
+
+    // A-F
+    attr.add(
+      widget is Container ? _widgetChild(widget.child) : '',
+    );
+
     // G-M
     attr.add(
-      widget is GestureDetector
-          ? _widgetChild(widget.child)
-          : widget is MouseRegion
-              ? _widgetChild(widget.child)
-              : widget is MultiChildRenderObjectWidget
-                  ? (widget is! RichText
-                      ? _widgetChildren(widget.children)
-                      : '')
-                  : '',
+      widget is GestureDetector ? _widgetChild(widget.child) : '',
     );
 
     // N-T
     attr.add(
       widget is ProxyWidget
           ? _widgetChild(widget.child)
-          : widget is SingleChildRenderObjectWidget
+          : widget is SingleChildScrollView
               ? _widgetChild(widget.child)
-              : widget is SingleChildScrollView
-                  ? _widgetChild(widget.child)
-                  : widget is SelectableText
-                      ? _inlineSpan(widget.textSpan!)
-                      : widget is Text
-                          ? widget.data!
-                          : widget is Tooltip
-                              ? _widgetChild(widget.child)
-                              : '',
+              : widget is SelectableText
+                  ? _inlineSpan(widget.textSpan!)
+                  : widget is Text
+                      ? widget.data!
+                      : widget is Tooltip
+                          ? _widgetChild(widget.child)
+                          : '',
     );
+
     // U-Z
+    attr.add('');
 
     final attrStr = attr.where((a) => a.isNotEmpty).join(',');
     return '[$type${attrStr.isNotEmpty ? ':$attrStr' : ''}]';
@@ -728,13 +734,13 @@ class HitTestApp extends StatelessWidget {
       );
 }
 
-class _RichTextFinder extends MatchFinder {
+class _TextFinder extends MatchFinder {
   final String data;
 
-  _RichTextFinder(this.data) : super(skipOffstage: true);
+  _TextFinder(this.data) : super(skipOffstage: true);
 
   @override
-  String get description => 'RichText "$data"';
+  String get description => '_TextFinder "$data"';
 
   @override
   bool matches(Element candidate) {
@@ -746,6 +752,16 @@ class _RichTextFinder extends MatchFinder {
         if (text.toPlainText() == data) {
           return true;
         }
+      }
+    } else if (widget is SelectableText) {
+      final text = widget.data;
+      if (text != null && text == data) {
+        return true;
+      }
+
+      final span = widget.textSpan;
+      if (span != null && span.toPlainText() == data) {
+        return true;
       }
     }
 
