@@ -246,6 +246,27 @@ void main() {
         ),
       );
     });
+
+    testWidgets('renders 100%', (WidgetTester tester) async {
+      const html = '<div style="min-height: 100%">Foo</div>';
+      final explained = await explain(
+        tester,
+        null,
+        hw: SingleChildScrollView(
+          child: HtmlWidget(
+            html,
+            key: hwKey,
+          ),
+        ),
+      );
+      expect(
+        explained,
+        equals(
+          '[CssSizing:height≥100.0%,width=100.0%,child='
+          '[RichText:(:Foo)]]',
+        ),
+      );
+    });
   });
 
   group('min-width', () {
@@ -303,6 +324,28 @@ void main() {
           '[CssSizing:width≥20.0,width=100.0%,child='
           '[RichText:(:Foo)]'
           ']',
+        ),
+      );
+    });
+
+    testWidgets('renders 100%', (WidgetTester tester) async {
+      const html = '<div style="min-width: 100%">Foo</div>';
+      final explained = await explain(
+        tester,
+        null,
+        hw: SingleChildScrollView(
+          scrollDirection: Axis.horizontal,
+          child: HtmlWidget(
+            html,
+            key: hwKey,
+          ),
+        ),
+      );
+      expect(
+        explained,
+        equals(
+          '[CssSizing:width≥100.0%,width=100.0%,child='
+          '[RichText:(:Foo)]]',
         ),
       );
     });
@@ -518,7 +561,13 @@ void main() {
       );
     });
 
-    final goldenSkip = Platform.isLinux ? null : 'Linux only';
+    final goldenSkipEnvVar = Platform.environment['GOLDEN_SKIP'];
+    final goldenSkip = goldenSkipEnvVar == null
+        ? Platform.isLinux
+            ? null
+            : 'Linux only'
+        : 'GOLDEN_SKIP=$goldenSkipEnvVar';
+
     GoldenToolkit.runWithConfiguration(
       () {
         group(
